volume: The volume is calculated by rounding.
When calculating the volume, we're going to take the integer part.
This may cause the left/right balance continued to change in one
direction when the volume was adjusted. So we adjusted to a rounded
way, accepting fine tuning of the left and right channels, but
remaining stable within the desired balance range.

Signed-off-by: lishengyu <lishengyu@uniontech.com>